work on ui

+++ src/api/category.ts
@@ -1,16 +1,23 @@
+'use server';
+
 import type { CategoryCreateType, CategoryType } from '@app/types/list.types';
+
+import { getUserToken } from './user';
 
 export async function createCategory(data: CategoryCreateType) {
   try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}categories`, {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-    return res.json();
+    const token = getUserToken();
+    if (token && token) {
+      const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}categories`, {
+        method: 'POST',
+        headers: {
+          Authorization: `Bearer ${token}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(data),
+      });
+      return await res.json();
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -18,18 +25,21 @@ export async function createCategory(data: CategoryCreateType) {
 
 export async function updateCategory(data: CategoryType) {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}categories/${data._id}`,
-      {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      }
-    );
-    return res.json();
+    const token = getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}categories/${data._id}`,
+        {
+          method: 'PUT',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(data),
+        }
+      );
+      return await res.json();
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -37,17 +47,20 @@ export async function updateCategory(data: CategoryType) {
 
 export async function deleteCategory(id: string) {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}categories/${id}`,
-      {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-      }
-    );
-    return res.json();
+    const token = getUserToken();
+    if (token) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}categories/${id}`,
+        {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${token}`,
+            'Content-Type': 'application/json',
+          },
+        }
+      );
+      return await res.json();
+    }
   } catch (e) {
     console.log('error', e);
   }
